refactor: move transactions list elements, logic, and styles to TransactionList component
